requirements.txt
python-dotenv
langchain
langchain_groq